Gaiadhi Chatbot/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
faiss-cpu==1.12.0
chromadb==0.5.5
tiktoken==0.7.0
openai==1.44.0
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.2
gTTS==2.5.4
elevenlabs==1.6.1
streamlit==1.38.0
gradio==4.44.0
plotly==5.17.0
requests==2.32.3
httpx==0.27.2
pydantic==2.9.2
sentence-transformers==3.0.1
scikit-learn==1.5.2

